python-common: HostSpec: add `validate()`
Adjust HostSpec interface to ServiceSpec
